scripts/requirements.txt
CairoSVG>=2.7.1,<3.0.0
Pillow>=9.5.0,<10.0.0 